secure_api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-telegram-bot>=21.0
cryptography>=41.0.0
